requirements.txt
# Core video & image processing
moviepy==1.0.3
Pillow>=8.0.0
imageio>=2.4.0
imageio-ffmpeg>=0.4.0
numpy>=1.19.0
requests>=2.25.0
fonttools>=4.0.0
scipy>=1.7.0

# YAML support
PyYAML

# Text-to-Speech dependencies
gtts
pydub
